refactor: Separate main module from __init__
