+++ jutge/utils.py
@@ -1,3 +1,16 @@
+from future.builtins import input
+
+
+def StdIn():
+    """
+    Generator that emulates `sys.stdin`. Uses `input()` builtin
+    rather than directly using sys.stdin to allow usage within an
+    interactive session.
+    """
+    while True:
+        yield input()
+
+
 def kwd_only(**kwd_ids_and_defaults):
     """For Python 2 compatibility (**kwargs must be used instead of
     keyword-only args). Decorates functions in such a way that only
